+++ source/icon.ts
@@ -4,6 +4,7 @@ export namespace Icons
     export let folder: vscode.IconPath;
     export let file: vscode.IconPath;
     export let error: vscode.IconPath;
+    export let star: vscode.IconPath;
     export let bookmark: vscode.IconPath;
     export let pin: vscode.IconPath;
     export let history: vscode.IconPath;
@@ -12,6 +13,11 @@ export namespace Icons
         folder = vscode.ThemeIcon.Folder;
         file = vscode.ThemeIcon.File;
         error = new vscode.ThemeIcon("error");
+        star =
+        {
+            light: vscode.Uri.joinPath(context.extensionUri, "images", "star.1024.svg"),
+            dark: vscode.Uri.joinPath(context.extensionUri, "images", "star-white.1024.svg"),
+        };
         bookmark =
         {
             light: vscode.Uri.joinPath(context.extensionUri, "images", "bookmark.1024.svg"),